Fix Derby pool.
Move all initilization code to new DerbyPooledDataSourceFactory.
