+++ droid-results/src/main/java/uk/gov/nationalarchives/droid/profile/datasource/DerbyPooledDataSource.java
@@ -31,7 +31,6 @@
  */
 package uk.gov.nationalarchives.droid.profile.datasource;
 
-import java.nio.file.Paths;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
@@ -42,8 +41,6 @@ import org.apache.commons.logging.LogFactory;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 
-import uk.gov.nationalarchives.droid.core.interfaces.config.RuntimeConfig;
-
 /**
  * @author rflitcroft
  *
@@ -51,11 +48,9 @@ import uk.gov.nationalarchives.droid.core.interfaces.config.RuntimeConfig;
 public class DerbyPooledDataSource extends HikariDataSource {
 
     private static final long serialVersionUID = -8613139738021279720L;
-    private static final String NO_CREATE_URL = "{none}";
 
     private final Log log = LogFactory.getLog(getClass());
     private final HikariConfig config;
-    private String createUrl = NO_CREATE_URL;
 
     /**
      * Constructor.
@@ -65,37 +60,6 @@ public class DerbyPooledDataSource extends HikariDataSource {
     public DerbyPooledDataSource(final HikariConfig config) {
         super(config);
         this.config = config;
-    }
-
-    /**
-     * Starts the database.
-     * @throws SQLException if the database could not be booted.
-     */
-    public void init() throws SQLException {
-        String droidLogDir = System.getProperty(RuntimeConfig.LOG_DIR);
-        System.setProperty("derby.stream.error.file", 
-                Paths.get(droidLogDir, "derby.log").toAbsolutePath().toString());
-        //setPoolPreparedStatements(true);
-        //setInitialSize(20); // initial size of connection pool.
-        log.debug(String.format("Booting database [%s]", config.getJdbcUrl()));
-        String url = getCreateURL();
-        String driverClassName = getDriverClassName();
-        try {
-            Class.forName(driverClassName);
-            DriverManager.getConnection(url).close();
-        } catch (ClassNotFoundException e) {
-            String message = String.format("Invalid driver class name: %s", driverClassName);
-            log.error(message, e);
-            throw new RuntimeException(message, e);
-        }
-    }
-    
-    private String getCreateURL() {
-        String url = config.getJdbcUrl() + ";create=true";
-        if (createUrl != null && !createUrl.isEmpty() && !NO_CREATE_URL.equals(createUrl)) {
-            url = url + ";" + createUrl;
-        }
-        return url;
     }
 
     /**
@@ -151,12 +115,5 @@ public class DerbyPooledDataSource extends HikariDataSource {
             log.error(e.getMessage(), e);
         }
     }
-     
-    /**
-     * 
-     * @param createUrl Sets the create Url to use when creating the database.
-     */
-    public void setCreateUrl(final String createUrl) {
-        this.createUrl = createUrl;
-    }
+
 }
